refactor: move waypoint to own package
