+++ backend/src/main/room/waypoint/waypoint_a_pojo.py
@@ -1,0 +1,10 @@
+from backend.src.main.game.values import DungeonCardValues
+from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
+
+
+class WaypointA(WaypointPOJO):
+    def __init__(self):
+        super(WaypointA, self).__init__(
+            DungeonCardValues.ENTRANCE_A,
+            DungeonCardValues.EXIT_A
+        )
