+++ backend/src/main/room/waypoint/waypoint_pojo.py
