+++ backend/src/main/tile/tile_geometry.py
@@ -2,7 +2,7 @@ from abc import ABC
 
 from backend.src.main.game.values import DungeonCardValues
 from backend.src.main.room.room import AbstractRoomCard
-from backend.src.main.room.waypoint_pojo import WaypointPOJO
+from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
 from backend.src.main.tile.center_tile import CenterTile
 from backend.src.main.tile.shift_tile import ShiftTile
