+++ backend/src/test/room/waypoint/test_waypoint_pojo.py
@@ -5,7 +5,7 @@ from backend.src.main.game.values import DungeonCardValues
 from backend.src.main.room.concrete_room_cards.burrow import Burrow
 from backend.src.main.room.concrete_room_cards.hovel import Hovel
 from backend.src.main.room.concrete_room_cards.trail import Trail
-from backend.src.main.room.waypoint_pojo import WaypointPOJO
+from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
 from backend.src.main.tile.tile import Tile
