+++ backend/src/test/tile/test_tile_geometry.py
@@ -5,7 +5,7 @@ import pytest
 from backend.src.main.game.values import DungeonCardValues
 from backend.src.main.room.concrete_room_cards.burrow import Burrow
 from backend.src.main.room.concrete_room_cards.tunnel import Tunnel
-from backend.src.main.room.waypoint_pojo import WaypointPOJO
+from backend.src.main.room.waypoint.waypoint_pojo import WaypointPOJO
 from backend.src.main.tile.center_tile import CenterTile
 from backend.src.main.tile.tile import Tile
 from backend.src.main.tile.tile_geometry import WaypointATileGeometry, WaypointBTileGeometry
